net: phy: AR40xx: convert to linkmode
Since kernel v5.4 old u32 supported and advertised features are converted to linkmode.
So lets convert it to linkmode in order to compile and work under kernel 5.4.

Also, since v5.4 setting features is a must have and they determine the capabilites of the PHY, so we must declare the features as gigabit PHY.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>