net: ethernet: ESSEDMA: convert to linkmode
Since kernel v5.4 old u32 supported and advertised features are converted to linkmode.

So lets convert it to linkmode in order to compile and work under kernel 5.4.

Also, old get/set settings API for ethtool is also dropped so lets convert to new ksettings API while at it as it demands linkmode.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>