net: AR40xx: fix MDIO get issue
MDIO drivers were moved into their own sub directory of networking drivers.
This has caused the AR40xx driver to probe before MDIO drivers and that wont work as it depends on the MDIO bus to be up so it can be fetched.

Lets solve it by moving the AR40xx into MDIO folder so they get probed like before.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>