net: dsa: fix calculation of CPU port stats offset
The CPU port statistics are placed after the statistics of the
master device, however this only works correctly if the master
netdev provides both the 'get_sset_count' and the 'get_ethtool_stats'
operations.

If the 'get_ethtool_ops' is missing, the offset remains zero,
and the CPU port statistics are stored into the place of the
master netdev's statistics.

  $ ethtool -S eth0
  NIC statistics:
       tx_q0_pkt: 5
       tx_q1_pkt: 0
       tx_q2_pkt: 30
       tx_q3_pkt: 0
       tx_q4_pkt: 0
       .
       .
       p00_RxBroad: 0
       p00_RxPause: 0
       p00_RxMulti: 0
       p00_RxFcsErr: 0
       p00_RxAlignErr: 0
       .
       .

Change the code to always get the count of the statistics
strings and compute the offset from that value.

  $ ethtool -S eth0
  NIC statistics:
       tx_q0_pkt: 0
       tx_q1_pkt: 0
       tx_q2_pkt: 0
       tx_q3_pkt: 0
       tx_q4_pkt: 0
       .
       .
       p00_RxBroad: 10
       p00_RxPause: 0
       p00_RxMulti: 31
       p00_RxFcsErr: 0
       p00_RxAlignErr: 0
       .
       .

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>