net: ipqess: use separate index for RX/TX rings and hardware queues
The driver uses the same index for the software rings and for
the hardware queues. This results in superfluous conversions in
the code where the ring indices are converted to hardware queue
indexes and vice versa. Add new field to the 'ipqess_tx_ring'
and 'ipqess_rx_ring' structures in preparation to avoid the
conversions.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>