net: ipqess: fix the address computation macros of RX queue registers
Some of the RX queue specific register address computation
macros are expecting the index of even numbered queues.
Although this works with the current code because the driver
uses the even numbered RX queues only but it makes it impossible
to use these macros for odd numbered queues.

Change the macros to compute the correct register address of
any queues, and set the queue index value in the RX ring to
the correct value to keep the existing behaviour.

Also remove the comment about this discrepancy, which is not
relevant anymore.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>