net: ipqess: fix svlan tag handling
The code shifts the svlan tag value to the upper 16 bits of
a 32 bits wide variable, then stores the value of this variable
into the 'svlan_tag' field of the TX descriptor. However the
field is 16 bits wide only, so it always get a zero value.

Remove the shift operator which is obviously wrong.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>