net: add support for DSA specific skb extensions
All of the already existing DSA tagging protocol drivers are storing
the tagging data directly into the skb. In most cases that is the only
way to send the required information to the underlying ethernet switch.

However on certain platforms (like the Qualcomm IPQ40xx SoCs) the
built-in ethernet switch is connected directly to an ethernet MAC,
and the tagging information must be sent out-of-band which is done
directly via the hardware TX descriptors of the ethernet MAC.

In such cases, putting the information into the skb causes unneccesary
overhead, because the ethernet driver must remove that before sending
the ethernet frame towards to the hardware.

This change makes it possible to send the tagging information
via skb extensions. With this approach, the twofold modifications
of the skb data can be avoided.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>