net: dsa: tag_ipq40xx: add new tagging protocol driver for IPQ40xx
This change adds a new tagging protocol driver for the built-in
ethernet switch of the Qualcomm Atheros IPQ40xx SoCs.

In comparison to the existing tagging protocols this hardware
requires a slightly different approach because the switch does
not use in-band tags.

On the receive path, the source port information is embedded
into the RX descriptors of the ethernet MAC hardware. Similarly,
the destination port mask must be sent via the TX descriptors
of the ethernet MAC when a packet is sent towards the switch.

In order to support this special requirements, this patch adds
a new tagging protocol driver.

The driver extracts the source port information directly from
the 'receive return descriptor' of the ethernet MAC. It is
possible because that descriptor is part of the skb received
from the ethernet driver.

Unfortunatley, it is not possible to put the destination port
information directly to the TX descriptors, because those are
handled internally by the driver of the ethernet hardware. To
overcome this limitation, the tagging driver uses the recently
introduced DSA specific SKB extension to send the destination
port information to the ethernet driver.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>