net: dsa: ipq40xx: only change the MIB_EN bit in MODULE_EN register
The MIB module needs to be enabled in the MODULE_EN register in
order to make it to counting. This is done in the qca8k_mib_init()
function. However instead of only changing the MIB module enable
bit, the function writes the whole register. As a side effect other
internal modules gets disabled.

Fix up the code to only change the MIB module specific bit.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>