net: dsa: ipq40xx: allow to compile QM error workaround code
It seems that the QM error workaround code is imported from the
ar40xx driver originally, however it does not compiled by default
because it is guarded by an ifdef.

It is questionable why the code is here at all if it is not even
compilable, but the change makes it buildable at least.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>