net: ipqess: ensure that TPDs has been written completely
The ipqess driver has severe issues with handling heavy network
traffic. This behaviour can be observed in various ways.

A simple ping test with large packets shows latency spikes:

  $ ping 192.168.1.157 -c 512 -A -s 65507
  PING 192.168.1.157 (192.168.1.157): 65507 data bytes
  ...
  65515 bytes from 192.168.1.157: seq=89 ttl=64 time=2.715 ms
  65515 bytes from 192.168.1.157: seq=91 ttl=64 time=20.090 ms
  65515 bytes from 192.168.1.157: seq=93 ttl=64 time=14.507 ms
  65515 bytes from 192.168.1.157: seq=94 ttl=64 time=2.738 ms
  ...
  65515 bytes from 192.168.1.157: seq=290 ttl=64 time=2.688 ms
  65515 bytes from 192.168.1.157: seq=291 ttl=64 time=20.988 ms
  65515 bytes from 192.168.1.157: seq=292 ttl=64 time=13.703 ms
  65515 bytes from 192.168.1.157: seq=293 ttl=64 time=2.480 ms

Sending large amount of TCP packets with iperf causes a stall
sooner or later:

  $ iperf -c 192.168.1.157 -n 5G -i 1
  ------------------------------------------------------------
  Client connecting to 192.168.1.157, TCP port 5001
  TCP window size:  105 KByte (default)
  ------------------------------------------------------------
  [  3] local 192.168.1.1 port 49702 connected with 192.168.1.157 port 5001
  [ ID] Interval       Transfer     Bandwidth
  ...
  [  3]  9.0-10.0 sec   110 MBytes   926 Mbits/sec
  [  3] 10.0-11.0 sec   101 MBytes   850 Mbits/sec
  [  3] 11.0-12.0 sec  63.6 KBytes   521 Kbits/sec
  [  3] 12.0-13.0 sec  0.00 Bytes  0.00 bits/sec

After this point the interface can't handle any further traffic.

While investigating the problem on the RX side, it is noticed
that the receiving interface has several RX frame errors which
are caused by too large packets:

  $ ifconfig  eth0 | grep RX
             RX packets:3672046 errors:8053 dropped:0 overruns:0 frame:8053
             RX bytes:1190314927 (1.1 GB)  TX bytes:21616 (21.6 KB)
  $ cat /sys/class/net/eth0/statistics/rx_length_errors
  8053

Although this is quite strange considering that the MTU value is
the same on both sides, but the error count is correct in fact.
The MIB counters of the respective switch port shows the same value:

  $ ethtool -S lan | grep -E 'Tx1518|TxMax'
       Tx1518Byte: 3590150
       TxMaxByte: 8053

Additionally, the MIB counters of the CPU port tells that as well:

  $ ethtool -S eth0 | grep -E 'Rx1518|RxMax'
       p00_Rx1518Byte: 3590150
       p00_RxMaxByte: 8053

This indicates that the large packets are coming to the switch's
CPU port from the EMAC. This should not happen given the fact that
the outgoing packet size is limited by the MTU of the interface
which is 1500 in this setup. The behaviour is the same even with
smaller MTU values, which leads to the conclusion that the large
packets are sent directly by the EDMA hardware.

It may happens either due to an internal hardware error or due
to  improper programming. The first case is unlikely, because
the hardware has no such issues with the original 'essedma'
driver.

Despite the different code of the two drivers, the hardware TX
descriptors are configured similarly so the problem should be
caused by something else. One thing which is possible that the
modifications in the TX descriptors are not yet seen by the
hardware when the software producer index register is updated.

This change adds a 'write memory barrier' statement right before
upgrading the hardware register, which ensures that all TPDs are
written completely before the hardware gets informed about the
new to be transmitted packets.

It is quite hard to say what happens inside the hardware, especially
without documentation so this fix might be wrong. However it cures
the problems described above, and the driver can pass large amount
data now.

  $ iperf -c 192.168.99.99 -i 1 -n 20G
  ------------------------------------------------------------
  Client connecting to 192.168.99.99, TCP port 5001
  TCP window size:  136 KByte (default)
  ------------------------------------------------------------
  [  3] local 192.168.99.1 port 39160 connected with 192.168.99.99 port 5001
  [ ID] Interval       Transfer     Bandwidth
  [  3]  0.0- 1.0 sec  98.5 MBytes   826 Mbits/sec
  [  3]  1.0- 2.0 sec   112 MBytes   940 Mbits/sec
  ...
  [  3] 181.0-182.0 sec   112 MBytes   936 Mbits/sec
  [  3] 182.0-183.0 sec   112 MBytes   940 Mbits/sec
  [  3]  0.0-183.1 sec  20.0 GBytes   938 Mbits/sec

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>