net: ath: Add support for ignoring EEPROM regulatory
Qualcomm requires hardcoded regulatory information into the
calibration/EEPROM data during manufacturing.

This is usually set to the World regulatory which is really strict,
and users cant enforce ones actually matching their countries
regulatory requirments provided by wireless-regdb as that can only
combine the default regulatory with the user selectable.

This is a big issue as it prevents a lot of channels in the 5GHz
band from being used at all, and more in the same band from being
used in AP mode.

This simply ignores the hardcoded regulatory data and uses the
user selected one.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>