net: dsa: ipq40xx: move regmap initialization to the probe function
The register mapping initialization happens in qca8k_setup()
currently. That function is only called during dsa switch
registration, so accessing the internal registers of the switch
are not accessible prior to that.

Move the regmap initialization code to the probe function in
order to allow accessing the registers earlier.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>