net: dsa: ipq40xx: don't use syscon based regmap access
At the moment the driver registers two regmap instances for accessing
the internal registers of the switch. One is a generic MMIO based,
and another one obtained via the syscon API.

This device is not a system controller at all so remove the syscon
based regmap and use the MMIO based one instead.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>