net: ipqess: remove VLAN tags from incoming packets
Via the NETIF_F_HW_VLAN_CTAG_RX flag, the driver indicates to
the network core that the hardware supports receive VLAN CTAG
acceleration. However the removal of the VLAN tags from the
incoming packets is not enabled in the hardware.

Change the code in order to configure the hardware correctly.
This allows to use VLAN interfaces on top of the interfaces
provided by the DSA layer. Although that it is working, but
the speed is horrible at the moment:

  $ ip -f inet addr show dev wan.19
  14: wan.19@wan: <BROADCAST,MULTICAST,UP,LOWER_UP> mtu 1500 qdisc noqueue state UP qlen 1000
      inet 10.10.19.1/24 brd 10.10.19.255 scope global wan.19
         valid_lft forever preferred_lft forever

  $ iperf -c 10.10.19.10 -i 1 -n 2G
  ------------------------------------------------------------
  Client connecting to 10.10.19.10, TCP port 5001
  TCP window size: 43.8 KByte (default)
  ------------------------------------------------------------
  [  3] local 10.10.19.1 port 54070 connected with 10.10.19.10 port 5001
  [ ID] Interval       Transfer     Bandwidth
  [  3]  0.0- 1.0 sec  2.75 MBytes  23.1 Mbits/sec
  [  3]  1.0- 2.0 sec  3.12 MBytes  26.2 Mbits/sec
  [  3]  2.0- 3.0 sec  2.75 MBytes  23.1 Mbits/sec
  [  3]  3.0- 4.0 sec  2.12 MBytes  17.8 Mbits/sec
  [  3]  4.0- 5.0 sec  3.25 MBytes  27.3 Mbits/sec
  [  3]  5.0- 6.0 sec  2.88 MBytes  24.1 Mbits/sec

This still needs further investigation.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>