net: dsa: tag_ipq40xx_ib: add in-band DSA tagging driver for IPQ40xx
This change adds another DSA tagging protocol driver for the built-in
switch of the IPQ40xx SoCs. In comparison to the already existing skb
extension based driver this protocol driver inserts a fake VLAN tag
into the outgoing packets.

The drawback is that the ethernet driver has to remove this tag before
it sends the packet to the EDMA hardware. It is required because the
destination port information must be passed via the tx descriptors to
the EDMA hardware.

Despite the overhead caused by the twofold vlan tag processing, this
driver gives an alternative of the skb extension based driver which
has performance problems on VLAN interfaces.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>