net: dsa: tag_8021q: avoid potential skb double free
The dsa_8021q_xmit() function is called by dsa_slave_xmit() which
frees the skb if NULL is returned. Also vlan_insert_tag() frees
the skb internally if an error happens. This potentially leads to
freeing the same skb twice.

To avoid this, use the __vlan_insert_tag() function which does
not frees the skb internally.

Compile tested only.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>