net: ipqess: fix TCP segmentation offload with VLAN tagged packets
The TCP segmentation offload feature does not work with
VLAN tagged packets. This can be observed during a simple
test done with iperf. While using the skb extension based
tagging driver, the transfer speed is horrible using a VLAN
interface on top of the DSA slave interface:

  $ iperf -c 10.10.19.10
  ------------------------------------------------------------
  Client connecting to 10.10.19.10, TCP port 5001
  TCP window size: 43.8 KByte (default)
  ------------------------------------------------------------
  [  3] local 10.10.19.1 port 52706 connected with 10.10.19.10 port 5001
  [ ID] Interval       Transfer     Bandwidth
  [  3]  0.0-10.0 sec  29.5 MBytes  24.7 Mbits/sec

Given the fact that the throughput is much better when using
the in-band tagging driver, the first logical conclusion
would be that the skb extension based tagging protocol driver
is broken. However that does not clarifies the fact, that using
the DSA slave interface directly has no such speed problems:

  $ iperf -c 192.168.99.99
  ------------------------------------------------------------
  Client connecting to 192.168.99.99, TCP port 5001
  TCP window size:  394 KByte (default)
  ------------------------------------------------------------
  [  3] local 192.168.99.1 port 47330 connected with 192.168.99.99 port 5001
  [ ID] Interval       Transfer     Bandwidth
  [  3]  0.0-10.0 sec  1.08 GBytes   931 Mbits/sec

Investigating the traffic on the receiving side shows broken
TCP headers. Another weird thing is that this only happens,
when the outgoing packet contains a VLAN tag.

Unfortunately, none of the existing and publicly available
IPQ40xx EDMA drivers contains any hint about how these kind
of traffix should be handled.

However there are several drivers are exist in linux for
different Qualcomm/Atheros ethernet devices and some of
those shows similarity to the EDMA hardware. What is more
interesting is that three of them are using a bit in the
Trasmit Packet Desctiptor for indicating that the outgoing
packet have a VLAN tag:

  $ grep -nr TAGGED drivers/net/ethernet/atheros/
  drivers/net/ethernet/atheros/atl1e/atl1e_main.c:1892: tpd->word3 |= 1 << TPD_VL_TAGGED_SHIFT;
  drivers/net/ethernet/atheros/atl1e/atl1e.h:139:#define TPD_VL_TAGGED_MASK 0x0001
  drivers/net/ethernet/atheros/atl1e/atl1e.h:140:#define TPD_VL_TAGGED_SHIFT 8
  drivers/net/ethernet/atheros/atlx/atl1.h:514:#define TPD_VL_TAGGED_MASK	0x0001
  drivers/net/ethernet/atheros/atlx/atl1.h:515:#define TPD_VL_TAGGED_SHIFT 8
  drivers/net/ethernet/atheros/alx/hw.h:131:#define TPD_VLTAGGED_MASK 0x0001
  drivers/net/ethernet/atheros/alx/hw.h:132:#define TPD_VLTAGGED_SHIFT 14
  drivers/net/ethernet/atheros/alx/hw.h:293:#define RRD_VLTAGGED_MASK 0x0001
  drivers/net/ethernet/atheros/alx/hw.h:294:#define RRD_VLTAGGED_SHIFT 16
  $

The bits used by the 'atl1e' and the 'atl1' drivers are
not applicable, but surprisingly it seems that the bit
used by the 'alx' driver has the same meaning in regard
to this EDMA hardware. Well, at least specifying that bit
in the TX descriptor when a VLAN tag is present fixes the
speed problem:

  $ iperf -c 10.10.19.10
  ------------------------------------------------------------
  Client connecting to 10.10.19.10, TCP port 5001
  TCP window size:  109 KByte (default)
  ------------------------------------------------------------
  [  3] local 10.10.19.1 port 52552 connected with 10.10.19.10 port 5001
  [ ID] Interval       Transfer     Bandwidth
  [  3]  0.0-10.0 sec  1.06 GBytes   914 Mbits/sec

This change adds definition for the VLAN_TAGGED bit, and
sets that bit in the TX descritor when the packet has a
VLAN tag.

Signed-off-by: Gabor Juhos <j4g8y7@gmail.com>