net: phylink: Treat speed change as a major change
Commit b7ad14c2fe2d ("net: phylink: re-implement interface configuration with
PCS") introduced a new check for a major change in link configuration.
If the change is not major mac_config will not be called.
This leads to the situation that speed setting for and an active link is ignored.
With this wa  mac_config will be called if the requested speed is different than
 the current link speed

Signed-off-by: Taras Chornyi <taras.chornyi@plvision.eu>