hwmon: Add Microchip PD69200 PoE PSE driver
Microchip PD69200  is a IEEE 802.3bt/at/af/PoH PoE controller.
It supports using UART or I2C for management, this driver only supports I2C.

It supports up to and including 48 ports (4-pair or 2-pair or any combination).

Exact feature support depends on the firmware version running on the PD69200,
the firmware can be updated.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>