gpio: tn48m: Add support for Delta TN4810M CPLD
Delta TN4810M uses a similar CPLD GPIO expander
like the TN48M, but it has pins for 48 SFP+ ports,
making a total of 192 pins.
It also provides the TX fault pins which the TN48M
does not.

Only TX disable pins like on the TN48M are output
ones.

Thankfully, regmap GPIO allows for the driver to be
easily extended to support the TN4810M.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>