net: dsa: qca8k: use actual number of ports in common code
Currently, qca8k uses the QCA8K_NUM_PORTS in the whole driver for the
operations requiring knowing number of ports it has.
This has worked fine so far as all of the supported switches have 7 ports,
however the IPQ4019 built-in variant of QCA8337N doesnt have the second
CPU port so it has 6 ports.

So, use the number of ports that is set to the DSA switch structure
instead of the define for the common code so that its correct for
functions requiring to know number of ports.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>