net: dsa: qca8k-ipq4019: dont flood ports
I previously set the driver to essentially flood all of the ports with
unknown traffic as it wouldn't work with unknown traffic being forwarded
to the CPU port.

This was as the Atheros header was not enabled and the switch was treating
all of that traffic as unknown and flooding the ports as configured.

So since the header is now properly enabled, don't just flood ports and
instead forward unknown traffic to CPU port for Linux to handle.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>