arm: dts: ipq4019: move PSGMII PHY to the switch node
Following up on the switch driver moving from getting the PSGMII PHY
register space as a syscon apply the required changes to the DT.

This allows getting rid of the PSGMII PHY as a separate DT node
completely as all of its resources are now under the switch node.
Currently only the register space is utilized, but this will most
likely change in the future.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>