arm64: dts: marvell: cn9130-crb: fix SD card detection
CN9130-CRB boards have a CD detect GPIO, but it looks like its just
connected to 3.3V and that the uSD card slot used doesnt even have a CD
tab/pin at all.
So, currently the GPIO will always remain high no matter whether there is
a uSD card inserted or not so removing and replugging the uSD card doesnt
work currently.

To work around that remove the CD GPIO and declare cd-broken so that
polling is used instead.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>