net: dsa: qca8k-ipq4019: use DSA assisted CPU learning
It looks like the HW based CPU port learning works fine for most use
cases, however it is currently breaking roaming.

The issue is that it will cause duplicate FDB entries and then you have
to wait for 5 minutes (The default aging time) or manually delete the
FDB entry.

So, until and if the HW based CPU port learning is fix lets utilize the
DSA provided assisted_learning_on_cpu_port and disable HW learning on
the CPU port.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>